Allow OCaml to prove Local.{index,level} distinct
OCaml cannot prove that `index` and `level` are distinct because they're abstract types, and indeed they were both defined as `int`. Exposing them as private single-variant types allows OCaml to prove them distinct so it can correctly infer impossible match cases over the type.

Since this is a breaking change in the interface of Context, the version number is bumped to 0.2.0.